refactor: share generation emitters and capture traversal
